Refactor phrase parsing and holidays into core modules
